requirements.txt
strands-agents
boto3
botocore
